some refactoring + readded incubator slot id
